versio 1.1.11

+++ src/main/java/modelo/Usuario.java
@@ -55,7 +55,7 @@ public class Usuario implements Serializable {
         this.tipoUsuario = tipoUsuario;
     }
 
-    public double getId() {
+    public int getId() {
         return id;
     }
